requirements.txt
numpy
scipy
bitarray
future
mockredispy
